[STABLE] Clean up mess in index.js
